refactor: add dom module
