+++ src/modules/dom.js
@@ -1,0 +1,386 @@
+const KEYS = [
+  [
+    {
+      className: 'Backquote',
+      eng: {
+        caps: '`',
+        caseDown: '`',
+        caseUp: '~',
+        shiftCaps: '~',
+      },
+      rus: {
+        caps: 'Ё',
+        caseDown: 'ё',
+        caseUp: 'Ё',
+        shiftCaps: 'ё',
+      },
+    },
+    {
+      className: 'Digit1',
+      eng: {
+        caps: '1',
+        caseDown: '1',
+        caseUp: '!',
+        shiftCaps: '!',
+      },
+      rus: {
+        caps: '1',
+        caseDown: '1',
+        caseUp: '!',
+        shiftCaps: '!',
+      },
+    },
+    {
+      className: 'Digit2',
+      eng: {
+        caps: '2',
+        caseDown: '2',
+        caseUp: '@',
+        shiftCaps: '@',
+      },
+      rus: {
+        caps: '2',
+        caseDown: '2',
+        caseUp: '"',
+        shiftCaps: '"',
+      },
+    },
+    {
+      className: 'Digit3',
+      eng: {
+        caps: '3',
+        caseDown: '3',
+        caseUp: '#',
+        shiftCaps: '#',
+      },
+      rus: {
+        caps: '3',
+        caseDown: '3',
+        caseUp: '№',
+        shiftCaps: '№',
+      },
+    },
+    {
+      className: 'Digit4',
+      eng: {
+        caps: '4',
+        caseDown: '4',
+        caseUp: '$',
+        shiftCaps: '$',
+      },
+      rus: {
+        caps: '4',
+        caseDown: '4',
+        caseUp: ';',
+        shiftCaps: ';',
+      },
+    },
+    {
+      className: 'Digit5',
+      eng: {
+        caps: '5',
+        caseDown: '5',
+        caseUp: '%',
+        shiftCaps: '%',
+      },
+      rus: {
+        caps: '5',
+        caseDown: '5',
+        caseUp: '%',
+        shiftCaps: '%',
+      },
+    },
+    {
+      className: 'Digit6',
+      eng: {
+        caps: '6',
+        caseDown: '6',
+        caseUp: '^',
+        shiftCaps: '^',
+      },
+      rus: {
+        caps: '6',
+        caseDown: '6',
+        caseUp: ':',
+        shiftCaps: ':',
+      },
+    },
+    {
+      className: 'Digit7',
+      eng: {
+        caps: '7',
+        caseDown: '7',
+        caseUp: '&',
+        shiftCaps: '&',
+      },
+      rus: {
+        caps: '7',
+        caseDown: '7',
+        caseUp: '?',
+        shiftCaps: '?',
+      },
+    },
+    {
+      className: 'Digit8',
+      eng: {
+        caps: '8',
+        caseDown: '8',
+        caseUp: '*',
+        shiftCaps: '*',
+      },
+      rus: {
+        caps: '8',
+        caseDown: '8',
+        caseUp: '*',
+        shiftCaps: '*',
+      },
+    },
+    {
+      className: 'Digit9',
+      eng: {
+        caps: '9',
+        caseDown: '9',
+        caseUp: '(',
+        shiftCaps: '(',
+      },
+      rus: {
+        caps: '9',
+        caseDown: '9',
+        caseUp: '(',
+        shiftCaps: '(',
+      },
+    },
+    {
+      className: 'Digit0',
+      eng: {
+        caps: '0',
+        caseDown: '0',
+        caseUp: ')',
+        shiftCaps: ')',
+      },
+      rus: {
+        caps: '0',
+        caseDown: '0',
+        caseUp: ')',
+        shiftCaps: ')',
+      },
+    },
+    {
+      className: 'Minus',
+      eng: {
+        caps: '-',
+        caseDown: '-',
+        caseUp: '_',
+        shiftCaps: '_',
+      },
+      rus: {
+        caps: '-',
+        caseDown: '-',
+        caseUp: '_',
+        shiftCaps: '_',
+      },
+    },
+    {
+      className: 'Equal',
+      eng: {
+        caps: '=',
+        caseDown: '=',
+        caseUp: '+',
+        shiftCaps: '+',
+      },
+      rus: {
+        caps: '=',
+        caseDown: '=',
+        caseUp: '+',
+        shiftCaps: '+',
+      },
+    },
+    { className: 'Backspace', eng: { caseDown: 'Backspace', caseUp: 'Backspace' }, rus: { caseDown: 'Backspace', caseUp: 'Backspace' } },
+  ],
+  [
+    { className: 'Tab', eng: { caseDown: 'Tab', caseUp: 'Tab' }, rus: { caseDown: 'Tab', caseUp: 'Tab' } },
+    { className: 'KeyQ', eng: { caseDown: 'q', caseUp: 'Q' }, rus: { caseDown: 'й', caseUp: 'Й' } },
+    { className: 'KeyW', eng: { caseDown: 'w', caseUp: 'W' }, rus: { caseDown: 'ц', caseUp: 'Ц' } },
+    { className: 'KeyE', eng: { caseDown: 'e', caseUp: 'E' }, rus: { caseDown: 'у', caseUp: 'У' } },
+    { className: 'KeyR', eng: { caseDown: 'r', caseUp: 'R' }, rus: { caseDown: 'к', caseUp: 'К' } },
+    { className: 'KeyT', eng: { caseDown: 't', caseUp: 'T' }, rus: { caseDown: 'е', caseUp: 'Е' } },
+    { className: 'KeyY', eng: { caseDown: 'y', caseUp: 'Y' }, rus: { caseDown: 'н', caseUp: 'Н' } },
+    { className: 'KeyU', eng: { caseDown: 'u', caseUp: 'U' }, rus: { caseDown: 'г', caseUp: 'Г' } },
+    { className: 'KeyI', eng: { caseDown: 'i', caseUp: 'I' }, rus: { caseDown: 'ш', caseUp: 'Ш' } },
+    { className: 'KeyO', eng: { caseDown: 'o', caseUp: 'O' }, rus: { caseDown: 'щ', caseUp: 'Щ' } },
+    { className: 'KeyP', eng: { caseDown: 'p', caseUp: 'P' }, rus: { caseDown: 'з', caseUp: 'З' } },
+    {
+      className: 'BracketLeft',
+      eng: {
+        caps: '[',
+        caseDown: '[',
+        caseUp: '{',
+        shiftCaps: '{',
+      },
+      rus: {
+        caps: 'Х',
+        caseDown: 'х',
+        caseUp: 'Х',
+        shiftCaps: 'х',
+      },
+    },
+    {
+      className: 'BracketRight',
+      eng: {
+        caps: ']',
+        caseDown: ']',
+        caseUp: '}',
+        shiftCaps: '}',
+      },
+      rus: {
+        caps: 'Ъ',
+        caseDown: 'ъ',
+        caseUp: 'Ъ',
+        shiftCaps: 'ъ',
+      },
+    },
+    {
+      className: 'Backslash',
+      eng: {
+        caps: '\\',
+        caseDown: '\\',
+        caseUp: '|',
+        shiftCaps: '|',
+      },
+      rus: {
+        caps: '\\',
+        caseDown: '\\',
+        caseUp: '/',
+        shiftCaps: '/',
+      },
+    },
+    { className: 'Delete', eng: { caseDown: 'Del', caseUp: 'Del' }, rus: { caseDown: 'Del', caseUp: 'Del' } },
+  ],
+  [
+    { className: 'CapsLock', eng: { caseDown: 'CapsLock', caseUp: 'CapsLock' }, rus: { caseDown: 'CapsLock', caseUp: 'CapsLock' } },
+    { className: 'KeyA', eng: { caseDown: 'a', caseUp: 'A' }, rus: { caseDown: 'ф', caseUp: 'Ф' } },
+    { className: 'KeyS', eng: { caseDown: 's', caseUp: 'S' }, rus: { caseDown: 'ы', caseUp: 'Ы' } },
+    { className: 'KeyD', eng: { caseDown: 'd', caseUp: 'D' }, rus: { caseDown: 'в', caseUp: 'В' } },
+    { className: 'KeyF', eng: { caseDown: 'f', caseUp: 'F' }, rus: { caseDown: 'а', caseUp: 'А' } },
+    { className: 'KeyG', eng: { caseDown: 'g', caseUp: 'G' }, rus: { caseDown: 'п', caseUp: 'П' } },
+    { className: 'KeyH', eng: { caseDown: 'h', caseUp: 'H' }, rus: { caseDown: 'р', caseUp: 'Р' } },
+    { className: 'KeyJ', eng: { caseDown: 'j', caseUp: 'J' }, rus: { caseDown: 'о', caseUp: 'О' } },
+    { className: 'KeyK', eng: { caseDown: 'k', caseUp: 'K' }, rus: { caseDown: 'л', caseUp: 'Л' } },
+    { className: 'KeyL', eng: { caseDown: 'l', caseUp: 'L' }, rus: { caseDown: 'д', caseUp: 'Д' } },
+    {
+      className: 'Semicolon',
+      eng: {
+        caps: ';',
+        caseDown: ';',
+        caseUp: ':',
+        shiftCaps: ':',
+      },
+      rus: {
+        caps: 'Ж',
+        caseDown: 'ж',
+        caseUp: 'Ж',
+        shiftCaps: 'ж',
+      },
+    },
+    {
+      className: 'Quote',
+      eng: {
+        caps: "'",
+        caseDown: "'",
+        caseUp: '"',
+        shiftCaps: '"',
+      },
+      rus: {
+        caps: 'Э',
+        caseDown: 'э',
+        caseUp: 'Э',
+        shiftCaps: 'э',
+      },
+    },
+    { className: 'Enter', eng: { caseDown: 'Enter', caseUp: 'Enter' }, rus: { caseDown: 'Enter', caseUp: 'Enter' } },
+  ],
+  [
+    { className: 'ShiftLeft', eng: { caseDown: 'Shift', caseUp: 'Shift' }, rus: { caseDown: 'Shift', caseUp: 'Shift' } },
+    { className: 'KeyZ', eng: { caseDown: 'z', caseUp: 'Z' }, rus: { caseDown: 'я', caseUp: 'Я' } },
+    { className: 'KeyX', eng: { caseDown: 'x', caseUp: 'X' }, rus: { caseDown: 'ч', caseUp: 'Ч' } },
+    { className: 'KeyC', eng: { caseDown: 'c', caseUp: 'C' }, rus: { caseDown: 'с', caseUp: 'С' } },
+    { className: 'KeyV', eng: { caseDown: 'v', caseUp: 'V' }, rus: { caseDown: 'м', caseUp: 'М' } },
+    { className: 'KeyB', eng: { caseDown: 'b', caseUp: 'B' }, rus: { caseDown: 'и', caseUp: 'И' } },
+    { className: 'KeyN', eng: { caseDown: 'n', caseUp: 'N' }, rus: { caseDown: 'т', caseUp: 'Т' } },
+    { className: 'KeyM', eng: { caseDown: 'm', caseUp: 'M' }, rus: { caseDown: 'ь', caseUp: 'Ь' } },
+    {
+      className: 'Comma',
+      eng: {
+        caps: ',',
+        caseDown: ',',
+        caseUp: '<',
+        shiftCaps: '<',
+      },
+      rus: {
+        caps: 'Б',
+        caseDown: 'б',
+        caseUp: 'Б',
+        shiftCaps: 'б',
+      },
+    },
+    {
+      className: 'Period',
+      eng: {
+        caps: '.',
+        caseDown: '.',
+        caseUp: '>',
+        shiftCaps: '>',
+      },
+      rus: {
+        caps: 'Ю',
+        caseDown: 'ю',
+        caseUp: 'Ю',
+        shiftCaps: 'ю',
+      },
+    },
+    {
+      className: 'Slash',
+      eng: {
+        caps: '/',
+        caseDown: '/',
+        caseUp: '?',
+        shiftCaps: '?',
+      },
+      rus: {
+        caps: '.',
+        caseDown: '.',
+        caseUp: ',',
+        shiftCaps: ',',
+      },
+    },
+    { className: 'ArrowUp', eng: { caseDown: '▲', caseUp: '▲' }, rus: { caseDown: '▲', caseUp: '▲' } },
+    { className: 'ShiftRight', eng: { caseDown: 'Shift', caseUp: 'Shift' }, rus: { caseDown: 'Shift', caseUp: 'Shift' } },
+  ],
+  [
+    { className: 'ControlLeft', eng: { caseDown: 'Ctrl', caseUp: 'Ctrl' }, rus: { caseDown: 'Ctrl', caseUp: 'Ctrl' } },
+    { className: 'MetaLeft', eng: { caseDown: 'Win', caseUp: 'Win' }, rus: { caseDown: 'Win', caseUp: 'Win' } },
+    { className: 'AltLeft', eng: { caseDown: 'Alt', caseUp: 'Alt' }, rus: { caseDown: 'Alt', caseUp: 'Alt' } },
+    { className: 'Space', eng: { caseDown: ' ', caseUp: ' ' }, rus: { caseDown: ' ', caseUp: ' ' } },
+    { className: 'AltRight', eng: { caseDown: 'Alt', caseUp: 'Alt' }, rus: { caseDown: 'Alt', caseUp: 'Alt' } },
+    { className: 'ArrowLeft', eng: { caseDown: '◄', caseUp: '◄' }, rus: { caseDown: '◄', caseUp: '◄' } },
+    { className: 'ArrowDown', eng: { caseDown: '▼', caseUp: '▼' }, rus: { caseDown: '▼', caseUp: '▼' } },
+    { className: 'ArrowRight', eng: { caseDown: '►', caseUp: '►' }, rus: { caseDown: '►', caseUp: '►' } },
+    { className: 'ControlRight', eng: { caseDown: 'Ctrl', caseUp: 'Ctrl' }, rus: { caseDown: 'Ctrl', caseUp: 'Ctrl' } },
+  ],
+];
+const SPECIALS = [
+  'Backspace',
+  'Tab',
+  'Enter',
+  'CapsLock',
+  'ShiftLeft',
+  'ShiftRight',
+  'AltLeft',
+  'AltRight',
+  'ControlLeft',
+  'ControlRight',
+  'MetaLeft',
+  'Delete',
+];
+
+export default { KEYS, SPECIALS };
